[refactor] SpringSecurity 수정에 따른 리팩토링
