+++ server/src/main/java/com/moovda_project/moovda/module/member/entity/Member.java
@@ -31,6 +31,9 @@ public class Member extends Auditable {
     @Column(length = 100, nullable = false)
     private String password;
 
+    @ElementCollection(fetch = FetchType.EAGER)
+    private List<String> roles = new ArrayList<>();
+
 //    @Column(nullable = false)
 //    private String nickname;
